requirements.txt
fastapi
uvicorn
boto3
pillow
torch
torchvision
numpy
python-multipart
python-dotenv